+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan03/hot100_random/day03/p005/Solution1047_02.java
@@ -1,6 +1,4 @@
-package com.noob.algorithm.daily.plan03.hot100_template.day03.p005;
-
-import java.util.Stack;
+package com.noob.algorithm.daily.plan03.hot100_random.day03.p005;
 
 /**
  * 🟢 1047 删除字符串中的所有相邻重复项 - https://leetcode.cn/problems/remove-all-adjacent-duplicates-in-string/description/
