+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan03/hot100_template/day03/p005/Solution1047_01.java
@@ -12,30 +12,7 @@ public class Solution1047_01 {
      * 思路分析：
      */
     public String removeDuplicates(String s) {
-        // 构建栈辅助存储
-        Stack<Character> stack = new Stack<>();
-        for (char ch : s.toCharArray()) {
-            // 校验栈顶元素与当前遍历元素
-            if (!stack.isEmpty()) {
-                char top = stack.peek();
-                if (top == ch) {
-                    // 弹出栈顶元素
-                    stack.pop();
-                } else {
-                    stack.push(ch);
-                }
-            } else {
-                stack.push(ch);
-            }
-        }
 
-        // 弹出栈内留存元素并拼接为字符串
-        StringBuffer buffer = new StringBuffer();
-        while (!stack.isEmpty()) {
-            buffer.append(stack.pop());
-        }
-
-        // 返回结果
-        return buffer.reverse().toString();
+        return null;
     }
 }
