+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan03/hot100_template/day03/p006/Solution226_01.java
@@ -14,24 +14,7 @@ public class Solution226_01 {
      * 思路分析：
      */
     public TreeNode invertTree(TreeNode root) {
-        dfs(root);
-        return root;
-    }
-
-    // 递归处理
-    private void dfs(TreeNode node) {
-        if (node == null) {
-            return;
-        }
-
-        // 翻转二叉树
-        TreeNode tmp = node.left;
-        node.left = node.right;
-        node.right = tmp;
-
-        // 递归处理
-        dfs(node.left);
-        dfs(node.right);
+        return null;
     }
 
 }
